% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_kernel.R
\docType{class}
\name{AbstractKernel-class}
\alias{AbstractKernel-class}
\title{Abstract Kernel Class}
\description{
This is an abstract class representing a generic kernel. It serves as a base class
for specific kernel implementations, providing a common interface for kernel operations.
}
