% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_kernel.R
\docType{class}
\name{ConstantKernel-class}
\alias{ConstantKernel-class}
\title{Constant Kernel Class}
\description{
A class representing a constant kernel, which inherits from the \code{AbstractKernel} class.
This kernel returns a constant value for all pairwise evaluations.
}
\section{Slots}{

\describe{
\item{\code{value_c}}{A numeric value representing the constant value of the kernel.}
}}

