% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distance-matrix.R
\name{diag_distance}
\alias{diag_distance}
\title{Compute a vector of point-aligned distances, vectorised when possible}
\usage{
diag_distance(distance_function, X1, X2)
}
\arguments{
\item{distance_function}{A distance function (see \link{distances}).}

\item{X1, X2}{Numeric matrices with the same number of rows.}
}
\value{
A numeric vector of length \code{nrow(X1)}.
}
\description{
Compute a vector of point-aligned distances, vectorised when possible
}
\keyword{internal}
