% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distance-matrix.R
\name{distance_matrix}
\alias{distance_matrix}
\title{Compute a full distance matrix, vectorised when possible}
\usage{
distance_matrix(distance_function, X1, X2)
}
\arguments{
\item{distance_function}{A distance function (see \link{distances}).}

\item{X1, X2}{Numeric matrices.}
}
\value{
A numeric matrix of dimension \code{nrow(X1)} x \code{nrow(X2)}.
}
\description{
Compute a full distance matrix, vectorised when possible
}
\keyword{internal}
