% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_kernel.R
\name{ensure_abstract_kernel}
\alias{ensure_abstract_kernel}
\title{Ensure an Object is an AbstractKernel}
\usage{
ensure_abstract_kernel(kernel)
}
\arguments{
\item{kernel}{An object to be checked or wrapped. If not an \code{AbstractKernel}, it should be a numeric value suitable for initializing a \code{ConstantKernel}.}
}
\value{
An object of class \code{AbstractKernel}.
}
\description{
This function checks if the provided object is an instance of the \code{AbstractKernel} class.
If not, it wraps the object in a \code{ConstantKernel}, provided the object is numeric.
}
\examples{
# Assuming ConstantKernel is a subclass of AbstractKernel
kernel <- ensure_abstract_kernel(2.5)
}
