% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_kernel.R
\name{get_hyperparameter_names}
\alias{get_hyperparameter_names}
\title{Get Hyperparameter Names}
\usage{
get_hyperparameter_names(kernel)
}
\arguments{
\item{kernel}{An object of class \code{AbstractKernel} or one of its subclasses.}
}
\value{
A character vector of hyperparameter names.
}
\description{
Retrieves the names of hyperparameters from a kernel object.
}
\examples{
# Assuming SEKernel is a subclass of AbstractKernel with hyperparameters
kernel <- new("SEKernel", variance_se = 1, length_scale_se = 1)
get_hyperparameter_names(kernel)
}
