% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_kernel.R
\name{get_hyperparameter_values}
\alias{get_hyperparameter_values}
\title{Get Hyperparameter Values}
\usage{
get_hyperparameter_values(kernel)
}
\arguments{
\item{kernel}{An object of class \code{AbstractKernel} or one of its subclasses.}
}
\value{
A named list of hyperparameter values.
}
\description{
Retrieves the values of hyperparameters from a kernel object.
}
\examples{
# Assuming SEKernel is a subclass of AbstractKernel with hyperparameters
kernel <- new("SEKernel", variance_se = 1, length_scale_se = 1)
get_hyperparameter_values(kernel)
}
