% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_kernel.R
\name{gt_HPs,AbstractKernel-method}
\alias{gt_HPs,AbstractKernel-method}
\title{Hyperparameters Method for AbstractKernel}
\usage{
\S4method{gt_HPs}{AbstractKernel}(obj)
}
\arguments{
\item{obj}{An object of class \code{AbstractKernel}.}
}
\description{
This method is intended to retrieve hyperparameters from an \code{AbstractKernel} object.
It must be implemented in subclasses.
}
