% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_kernel.R
\name{gt_HPs,ConstantKernel-method}
\alias{gt_HPs,ConstantKernel-method}
\title{Hyperparameters Method for ConstantKernel}
\usage{
\S4method{gt_HPs}{ConstantKernel}(obj)
}
\arguments{
\item{obj}{An object of class \code{ConstantKernel}.}
}
\value{
A list containing the constant value of the kernel.
}
\description{
Retrieves the hyperparameters of a \code{ConstantKernel} object.
}
