% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_kernel.R
\name{gt_HPs,LinearKernel-method}
\alias{gt_HPs,LinearKernel-method}
\title{Hyperparameters Method for LinearKernel}
\usage{
\S4method{gt_HPs}{LinearKernel}(obj)
}
\arguments{
\item{obj}{An object of class \code{LinearKernel}.}
}
\value{
A list containing the bias variance, variance scaling factor, and constant offset.
}
\description{
Retrieves the hyperparameters of a \code{LinearKernel} object.
}
