% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_kernel.R
\name{gt_HPs,MaternKernel12-method}
\alias{gt_HPs,MaternKernel12-method}
\title{Hyperparameters Method for MaternKernel12}
\usage{
\S4method{gt_HPs}{MaternKernel12}(obj)
}
\arguments{
\item{obj}{An object of class \code{MaternKernel12}.}
}
\value{
A list containing the length scale parameter.
}
\description{
Retrieves the hyperparameters of a \code{MaternKernel12} object.
}
