% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_kernel.R
\name{gt_HPs,NoiseKernel-method}
\alias{gt_HPs,NoiseKernel-method}
\title{Hyperparameters Method for NoiseKernel}
\usage{
\S4method{gt_HPs}{NoiseKernel}(obj)
}
\arguments{
\item{obj}{An object of class \code{NoiseKernel}.}
}
\value{
A list containing the noise variance value.
}
\description{
Retrieves the hyperparameters of a \code{NoiseKernel} object.
}
