% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_kernel.R
\name{gt_HPs,PeriodicKernel-method}
\alias{gt_HPs,PeriodicKernel-method}
\title{Hyperparameters Method for PeriodicKernel}
\usage{
\S4method{gt_HPs}{PeriodicKernel}(obj)
}
\arguments{
\item{obj}{An object of class \code{PeriodicKernel}.}
}
\value{
A list containing the variance, length scale, and period parameters.
}
\description{
Retrieves the hyperparameters of a \code{PeriodicKernel} object.
}
