% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_kernel.R
\name{gt_HPs,ProductKernel-method}
\alias{gt_HPs,ProductKernel-method}
\title{Hyperparameters Method for ProductKernel}
\usage{
\S4method{gt_HPs}{ProductKernel}(obj)
}
\arguments{
\item{obj}{An object of class \code{ProductKernel}.}
}
\value{
A list of hyperparameters from all constituent kernels.
}
\description{
Retrieves the hyperparameters of a \code{ProductKernel} object by combining the hyperparameters of its constituent kernels.
}
