% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_kernel.R
\name{gt_HPs,RationalQuadraticKernel-method}
\alias{gt_HPs,RationalQuadraticKernel-method}
\title{Hyperparameters Method for RationalQuadraticKernel}
\usage{
\S4method{gt_HPs}{RationalQuadraticKernel}(obj)
}
\arguments{
\item{obj}{An object of class \code{RationalQuadraticKernel}.}
}
\value{
A list containing the variance, length scale, and alpha parameters.
}
\description{
Retrieves the hyperparameters of a \code{RationalQuadraticKernel} object.
}
