% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_kernel.R
\name{gt_HPs,SEKernel-method}
\alias{gt_HPs,SEKernel-method}
\title{Hyperparameters Method for SEKernel}
\usage{
\S4method{gt_HPs}{SEKernel}(obj)
}
\arguments{
\item{obj}{An object of class \code{SEKernel}.}
}
\value{
A list containing the variance and length scale parameters.
}
\description{
Retrieves the hyperparameters of an \code{SEKernel} object.
}
