% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_kernel.R
\name{gt_HPs,SumKernel-method}
\alias{gt_HPs,SumKernel-method}
\title{Hyperparameters Method for SumKernel}
\usage{
\S4method{gt_HPs}{SumKernel}(obj)
}
\arguments{
\item{obj}{An object of class \code{SumKernel}.}
}
\value{
A list of hyperparameters from all constituent kernels.
}
\description{
Retrieves the hyperparameters of a \code{SumKernel} object by combining the hyperparameters of its constituent kernels.
}
