% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_kernel.R
\name{gt_HPs}
\alias{gt_HPs}
\title{Generic Function for Getting Hyperparameters}
\usage{
gt_HPs(obj)
}
\arguments{
\item{obj}{An object inheriting from the \code{AbstractKernel} class.}
}
\value{
A list of hyperparameters.
}
\description{
This generic function retrieves the hyperparameters of a kernel object.
}
