% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_kernel.R
\name{initialize,ConstantKernel-method}
\alias{initialize,ConstantKernel-method}
\title{Initialize Method for ConstantKernel}
\usage{
\S4method{initialize}{ConstantKernel}(.Object, value_c = 1)
}
\arguments{
\item{.Object}{An object of class \code{ConstantKernel}.}

\item{value_c}{A numeric value for the constant kernel. Defaults to a random uniform value between 0 and 3.}
}
\value{
An initialized object of class \code{ConstantKernel}.
}
\description{
Initializes an instance of the \code{ConstantKernel} class with a specified constant value.
}
