% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_kernel.R
\name{initialize,LinearKernel-method}
\alias{initialize,LinearKernel-method}
\title{Initialize Method for LinearKernel}
\usage{
\S4method{initialize}{LinearKernel}(.Object, sigma2_b = 1, sigma2_v = 1, c = 1)
}
\arguments{
\item{.Object}{An object of class \code{LinearKernel}.}

\item{sigma2_b}{A numeric value for the bias variance. Defaults to a random uniform value between 0 and 3.}

\item{sigma2_v}{A numeric value for the variance scaling factor. Defaults to a random uniform value between 0 and 3.}

\item{c}{A numeric value for the constant offset. Defaults to a random uniform value between 0 and 3.}
}
\value{
An initialized object of class \code{LinearKernel}.
}
\description{
Initializes an instance of the \code{LinearKernel} class with specified bias variance, variance scaling factor, and constant offset.
}
