% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_kernel.R
\name{initialize,MaternKernel52-method}
\alias{initialize,MaternKernel52-method}
\title{Initialize Method for MaternKernel52}
\usage{
\S4method{initialize}{MaternKernel52}(.Object, length_scale_mat = 1)
}
\arguments{
\item{.Object}{An object of class \code{MaternKernel52}.}

\item{length_scale_mat}{A numeric value for the length scale parameter.}
}
\value{
An initialized object of class \code{MaternKernel52}.
}
\description{
Initializes an instance of the \code{MaternKernel52} class with a specified length scale parameter.
}
