% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_kernel.R
\name{initialize,NoiseKernel-method}
\alias{initialize,NoiseKernel-method}
\title{Initialize Method for NoiseKernel}
\usage{
\S4method{initialize}{NoiseKernel}(.Object, value_c = 1)
}
\arguments{
\item{.Object}{An object of class \code{NoiseKernel}.}

\item{value_c}{A numeric value for the noise variance. Defaults to a random uniform value between 0 and 3.}
}
\value{
An initialized object of class \code{NoiseKernel}.
}
\description{
Initializes an instance of the \code{NoiseKernel} class with a specified noise variance value.
}
