% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_kernel.R
\name{initialize,PeriodicKernel-method}
\alias{initialize,PeriodicKernel-method}
\title{Initialize Method for PeriodicKernel}
\usage{
\S4method{initialize}{PeriodicKernel}(.Object, variance_per = 1, length_scale_per = 1, period = 1)
}
\arguments{
\item{.Object}{An object of class \code{PeriodicKernel}.}

\item{variance_per}{A numeric value for the signal variance.}

\item{length_scale_per}{A numeric value for the length scale parameter.}

\item{period}{A numeric value for the period.}
}
\value{
An initialized object of class \code{PeriodicKernel}.
}
\description{
Initializes an instance of the \code{PeriodicKernel} class with specified variance, length scale, and period parameters.
}
