% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_kernel.R
\name{initialize,ProductKernel-method}
\alias{initialize,ProductKernel-method}
\title{Initialize Method for ProductKernel}
\usage{
\S4method{initialize}{ProductKernel}(.Object, kernels)
}
\arguments{
\item{.Object}{An object of class \code{ProductKernel}.}

\item{kernels}{A list of kernel objects to be multiplied.}
}
\value{
An initialized object of class \code{ProductKernel}.
}
\description{
Initializes an instance of the \code{ProductKernel} class with a list of kernel objects.
}
