% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_kernel.R
\name{initialize,RationalQuadraticKernel-method}
\alias{initialize,RationalQuadraticKernel-method}
\title{Initialize Method for RationalQuadraticKernel}
\usage{
\S4method{initialize}{RationalQuadraticKernel}(.Object, variance_rq = 1, length_scale_rq = 1, alpha_rq = 0.1)
}
\arguments{
\item{.Object}{An object of class \code{RationalQuadraticKernel}.}

\item{variance_rq}{A numeric value for the signal variance. Defaults to a random uniform value between 0 and 3.}

\item{length_scale_rq}{A numeric value for the length scale parameter. Defaults to a random uniform value between 0 and 3.}

\item{alpha_rq}{A numeric value for the scale mixture parameter. Defaults to a random uniform value between 0 and 3.}
}
\value{
An initialized object of class \code{RationalQuadraticKernel}.
}
\description{
Initializes an instance of the \code{RationalQuadraticKernel} class with specified variance, length scale, and alpha parameters.
}
