% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_kernel.R
\name{initialize,SEKernel-method}
\alias{initialize,SEKernel-method}
\title{Initialize Method for SEKernel}
\usage{
\S4method{initialize}{SEKernel}(.Object, variance_se = 1, length_scale_se = 1)
}
\arguments{
\item{.Object}{An object of class \code{SEKernel}.}

\item{variance_se}{A numeric value for the signal variance. Defaults to a random uniform value between 0 and 3.}

\item{length_scale_se}{A numeric value for the length scale parameter. Defaults to a random uniform value between 0 and 3.}
}
\value{
An initialized object of class \code{SEKernel}.
}
\description{
Initializes an instance of the \code{SEKernel} class with specified variance and length scale parameters.
}
