% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_kernel.R
\name{initialize,SumKernel-method}
\alias{initialize,SumKernel-method}
\title{Initialize Method for SumKernel}
\usage{
\S4method{initialize}{SumKernel}(.Object, kernels)
}
\arguments{
\item{.Object}{An object of class \code{SumKernel}.}

\item{kernels}{A list of kernel objects to be summed.}
}
\value{
An initialized object of class \code{SumKernel}.
}
\description{
Initializes an instance of the \code{SumKernel} class with a list of kernel objects.
}
