% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernel.R
\name{kde}
\alias{kde}
\title{Kernel density estimation based on observation.}
\usage{
kde(x, Xi)
}
\arguments{
\item{x}{Setting of Domain for kernel density estimation}

\item{Xi}{Observation}
}
\description{
This function generate pdf data for kernel density 
estimation. Kernel is simply the N(0,1).
}
\examples{
Xi <- c(3,4.5,5,8,9)
x <- seq(min(Xi)-3, max(Xi)+3,by=0.1)
sx <- seq(-3,3,by=0.1)

haty <- kde(x,Xi)

plot(x,haty,type="l",ylim=c(0,max(haty)))
for(i in 1:length(Xi)){
   lines(sx+Xi[i], dnorm(sx)/length(Xi),lty=2) 
}
}
\keyword{kernel}
