% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_kernel.R
\name{kernel_deriv,AbstractKernel-method}
\alias{kernel_deriv,AbstractKernel-method}
\title{Kernel Derivative Method for AbstractKernel}
\usage{
\S4method{kernel_deriv}{AbstractKernel}(obj, x, y, param)
}
\arguments{
\item{obj}{An object of class \code{AbstractKernel}.}

\item{x}{A matrix of input vectors.}

\item{y}{A matrix of input vectors.}

\item{param}{The parameter with respect to which the derivative is computed.}
}
\description{
This method computes the derivative of the kernel function with respect to a parameter.
It must be implemented in subclasses.
}
