% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_kernel.R
\name{kernel_deriv,ConstantKernel-method}
\alias{kernel_deriv,ConstantKernel-method}
\title{Kernel Derivative Method for ConstantKernel}
\usage{
\S4method{kernel_deriv}{ConstantKernel}(obj, x, y, param)
}
\arguments{
\item{obj}{An object of class \code{ConstantKernel}.}

\item{x}{A matrix of input vectors.}

\item{y}{A matrix of input vectors.}

\item{param}{The parameter with respect to which the derivative is computed.}
}
\value{
A matrix representing the derivative of the kernel evaluations.
}
\description{
Computes the derivative of the kernel function with respect to a parameter for a \code{ConstantKernel} object.
The derivative with respect to the constant value is a matrix of ones.

The derivative of \eqn{K_{\text{cst}}(x, x')} with respect to \eqn{c} is given by:

\eqn{\frac{dK_{\text{cst}}(x, x')}{dc} = \begin{bmatrix}
1 & 1 & \cdots & 1 \\
1 & 1 & \cdots & 1 \\
\vdots & \vdots & \ddots & \vdots \\
1 & 1 & \cdots & 1
\end{bmatrix}}
}
