% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_kernel.R
\name{kernel_deriv,LinearKernel-method}
\alias{kernel_deriv,LinearKernel-method}
\title{Kernel Derivative Method for LinearKernel}
\usage{
\S4method{kernel_deriv}{LinearKernel}(obj, x, y, param)
}
\arguments{
\item{obj}{An object of class \code{LinearKernel}.}

\item{x}{A matrix of input vectors.}

\item{y}{A matrix of input vectors.}

\item{param}{The parameter with respect to which the derivative is computed. Can be "sigma2_b", "sigma2_v", or "c".}
}
\value{
A matrix representing the derivative of the kernel evaluations.
}
\description{
Computes the derivative of the kernel function with respect to a parameter for a \code{LinearKernel} object.
Supports derivatives with respect to the bias variance, variance scaling factor, and constant offset.

The derivative of \eqn{K_{\text{Lin}}(x, x')} with respect to \eqn{\sigma^2_b} is given by:

\eqn{\frac{dK_{\text{Lin}}(x, x')}{d\sigma^2_b} = 1}

The derivative of \eqn{K_{\text{Lin}}(x, x')} with respect to \eqn{\sigma^2_v} is given by:

\eqn{\frac{dK_{\text{Lin}}(x, x')}{d\sigma^2_v} = (x - c)(x' - c)}

\eqn{\frac{dK_{\text{Lin}}(x, x')}{d\sigma^2_v} = \frac{K_{\text{Lin}}(x, x')}{\sigma^2_v}}

The derivative of \eqn{K_{\text{Lin}}(x, x')} with respect to \eqn{c} is given by:

\eqn{\frac{dK_{\text{Lin}}(x, x')}{dc} = \sigma^2_v \left( 2c - x - x' \right)}

\eqn{\frac{dK_{\text{Lin}}(x, x')}{dc} = -\sigma^2_v \left( (x - c) + (x' - c) \right)}
}
