% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_kernel.R
\name{kernel_deriv,MaternKernel12-method}
\alias{kernel_deriv,MaternKernel12-method}
\title{Kernel Derivative Method for MaternKernel12}
\usage{
\S4method{kernel_deriv}{MaternKernel12}(obj, x, y, param)
}
\arguments{
\item{obj}{An object of class \code{MaternKernel12}.}

\item{x}{A matrix of input vectors.}

\item{y}{A matrix of input vectors.}

\item{param}{The parameter with respect to which the derivative is computed. Should be "length_scale_mat".}
}
\value{
A matrix representing the derivative of the kernel evaluations.
}
\description{
Computes the derivative of the kernel function with respect to the length scale parameter for a \code{MaternKernel12} object.

The derivative of \eqn{K_{\text{Mat}_{\nu = \frac{1}{2}}}(x, x')} with respect to \eqn{\sigma^2} is given by:

\eqn{\frac{dK_{\text{Mat}_{\nu = \frac{1}{2}}}(x, x')}{d\sigma^2} = \exp\left(-\frac{\|x - x'\|^2}{\ell}\right)}

\eqn{\frac{dK_{\text{Mat}_{\nu = \frac{1}{2}}}(x, x')}{d\sigma^2} = \frac{K_{\text{Mat}_{\nu = \frac{1}{2}}}(x, x')}{\sigma^2}}

The derivative of \eqn{K_{\text{Mat}_{\nu = \frac{1}{2}}}(x, x')} with respect to \eqn{\ell} is given by:

\eqn{\frac{dK_{\text{Mat}_{\nu = \frac{1}{2}}}(x, x')}{d\ell} = -\exp\left(-\frac{\|x - x'\|^2}{\ell}\right) \cdot \|x - x'\|^2 \cdot (-\ell^{-2})}

\eqn{\frac{dK_{\text{Mat}_{\nu = \frac{1}{2}}}(x, x')}{d\ell} = \frac{\|x - x'\|^2}{\ell^2} \cdot K_{\text{Mat}_{\nu = \frac{1}{2}}}(x, x')}
}
