% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_kernel.R
\name{kernel_deriv,MaternKernel32-method}
\alias{kernel_deriv,MaternKernel32-method}
\title{Kernel Derivative Method for MaternKernel32}
\usage{
\S4method{kernel_deriv}{MaternKernel32}(obj, x, y, param)
}
\arguments{
\item{obj}{An object of class \code{MaternKernel32}.}

\item{x}{A matrix of input vectors.}

\item{y}{A matrix of input vectors.}

\item{param}{The parameter with respect to which the derivative is computed. Should be "length_scale_mat".}
}
\value{
A matrix representing the derivative of the kernel evaluations.
}
\description{
Computes the derivative of the kernel function with respect to the length scale parameter for a \code{MaternKernel32} object.

The derivative of \eqn{K_{\text{Mat}_{\nu = \frac{3}{2}}}(x, x')} with respect to \eqn{\sigma^2} is given by:

\eqn{\frac{dK_{\text{Mat}_{\nu = \frac{3}{2}}}(x, x')}{d\sigma^2} = \left( \frac{\sqrt{3} \cdot \|x - x'\|^2}{\ell} + 1 \right) \exp\left(-\frac{\sqrt{3} \cdot \|x - x'\|^2}{\ell}\right)}

\eqn{\frac{dK_{\text{Mat}_{\nu = \frac{3}{2}}}(x, x')}{d\sigma^2} = \frac{K_{\text{Mat}_{\nu = \frac{3}{2}}}(x, x')}{\sigma^2}}

The derivative of \eqn{K_{\text{Mat}_{\nu = \frac{3}{2}}}(x, x')} with respect to \eqn{\ell} is given by:

\eqn{\frac{dK_{\text{Mat}_{\nu = \frac{3}{2}}}(x, x')}{d\ell} = \frac{3 \cdot (\|x - x'\|^2)^2 \cdot \exp\left(-\frac{\sqrt{3} \|x - x'\|^2}{\ell}\right)}{\ell^3}}
}
