% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_kernel.R
\name{kernel_deriv,NoiseKernel-method}
\alias{kernel_deriv,NoiseKernel-method}
\title{Kernel Derivative Method for NoiseKernel}
\usage{
\S4method{kernel_deriv}{NoiseKernel}(obj, x, y, param)
}
\arguments{
\item{obj}{An object of class \code{NoiseKernel}.}

\item{x}{A matrix of input vectors.}

\item{y}{A matrix of input vectors.}

\item{param}{The parameter with respect to which the derivative is computed.}
}
\value{
A diagonal matrix representing the derivative of the kernel evaluations.
}
\description{
Computes the derivative of the kernel function with respect to a parameter for a \code{NoiseKernel} object.
The derivative with respect to the noise variance is a diagonal matrix of ones.

The derivative of \eqn{K_{\text{noise}}(x, x')} with respect to \eqn{c} is given by:

\eqn{\frac{dK_{\text{noise}}(x, x')}{dc} = I_n = \begin{bmatrix}
1 & 0 & \cdots & 0 \\
0 & 1 & \cdots & 0 \\
\vdots & \vdots & \ddots & \vdots \\
0 & 0 & \cdots & 1
\end{bmatrix}}
}
