% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_kernel.R
\name{kernel_deriv,PeriodicKernel-method}
\alias{kernel_deriv,PeriodicKernel-method}
\title{Kernel Derivative Method for PeriodicKernel}
\usage{
\S4method{kernel_deriv}{PeriodicKernel}(obj, x, y, param)
}
\arguments{
\item{obj}{An object of class \code{PeriodicKernel}.}

\item{x}{A matrix of input vectors.}

\item{y}{A matrix of input vectors.}

\item{param}{The parameter with respect to which the derivative is computed. Can be "variance_per", "length_scale_per", or "period".}
}
\value{
A matrix representing the derivative of the kernel evaluations.
}
\description{
Computes the derivative of the kernel function with respect to a parameter for a \code{PeriodicKernel} object.
Supports derivatives with respect to the variance, length scale, and period parameters.

The derivative of \eqn{K_{\text{Per}}(x, x')} with respect to \eqn{\sigma^2} is given by:

\eqn{\frac{dK_{\text{Per}}(x, x')}{d\sigma^2} = \exp \left( -\frac{2}{\ell^2} \sin^2 \left( \pi \frac{x - x'}{p} \right) \right)}

\eqn{\frac{dK_{\text{Per}}(x, x')}{d\sigma^2} = \frac{K_{\text{Per}}(x, x')}{\sigma^2}}

The derivative of \eqn{K_{\text{Per}}(x, x')} with respect to \eqn{\ell} is given by:

\eqn{\frac{dK_{\text{Per}}(x, x')}{d\ell} = \frac{4\sigma^2 \sin^2\left(\frac{\pi(x-x')}{p}\right) \exp\left(-\frac{2\sin^2\left(\frac{\pi(x-x')}{p}\right)}{\ell^2}\right)}{\ell^3}}

The derivative of \eqn{K_{\text{Per}}(x, x')} with respect to \eqn{p} is given by:

\eqn{\frac{dK_{\text{Per}}(x, x')}{dp} = \frac{4\pi\sigma^2(x-x')\sin\left(\frac{\pi(x-x')}{p}\right)\cos\left(\frac{\pi(x-x')}{p}\right)\exp\left(-\frac{2\sin^2\left(\frac{\pi(x-x')}{p}\right)}{\ell^2}\right)}{\ell^2 p^2}}

\eqn{\frac{dK_{\text{Per}}(x, x')}{dp} = \frac{2\pi\sigma^2(x-x')\sin\left(\frac{2\pi(x-x')}{p}\right)\exp\left(-\frac{2\sin^2\left(\frac{\pi(x-x')}{p}\right)}{\ell^2}\right)}{\ell^2 p^2}}
}
