% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_kernel.R
\name{kernel_deriv,RationalQuadraticKernel-method}
\alias{kernel_deriv,RationalQuadraticKernel-method}
\title{Kernel Derivative Method for RationalQuadraticKernel}
\usage{
\S4method{kernel_deriv}{RationalQuadraticKernel}(obj, x, y, param)
}
\arguments{
\item{obj}{An object of class \code{RationalQuadraticKernel}.}

\item{x}{A matrix of input vectors.}

\item{y}{A matrix of input vectors.}

\item{param}{The parameter with respect to which the derivative is computed. Can be "variance_rq", "length_scale_rq", or "alpha_rq".}
}
\value{
A matrix representing the derivative of the kernel evaluations.
}
\description{
Computes the derivative of the kernel function with respect to a parameter for a \code{RationalQuadraticKernel} object.
Supports derivatives with respect to the variance, length scale, and alpha parameters.

The derivative of \eqn{K_{\text{RQ}}(x, x')} with respect to \eqn{\sigma^2} is given by:

\eqn{\frac{dK_{\text{RQ}}(x, x')}{d\sigma^2} = \left(1 + \frac{\|x - x'\|^2}{2 \alpha \ell^2}\right)^{-\alpha}}

\eqn{\frac{dK_{\text{RQ}}(x, x')}{d\sigma^2} = \frac{K_{\text{RQ}}(x, x')}{\sigma^2}}

The derivative of \eqn{K_{\text{RQ}}(x, x')} with respect to \eqn{\alpha} is given by:

\eqn{\frac{dK_{\text{RQ}}(x, x')}{d\alpha} = \sigma^2 \left( \frac{\|x - x'\|^2}{2 \alpha \ell^2} + 1 \right)^{-\alpha} \left( \frac{\|x - x'\|^2}{2\alpha \ell^2} \left( \frac{\|x - x'\|^2}{2 \alpha \ell^2} + 1 \right)^{-1} - \ln \left( \frac{\|x - x'\|^2}{2\alpha \ell^2} + 1 \right) \right)}

\eqn{\frac{dK_{\text{RQ}}(x, x')}{d\alpha} = -\sigma^2 \left( \frac{\|x - x'\|^2}{2 \alpha \ell^2} + 1 \right)^{-\alpha} \left( \frac{(2 \alpha \ell^2 + \|x - x'\|^2) \ln \left( \frac{\|x - x'\|^2}{2 \alpha \ell^2} + 1 \right) - \|x - x'\|^2}{2 \alpha \ell^2 + \|x - x'\|^2} \right)}

\eqn{\frac{dK_{\text{RQ}}(x, x')}{d\alpha} = -K_{\text{RQ}}(x, x') \left( \frac{(2 \alpha \ell^2 + \|x - x'\|^2) \ln \left( \frac{\|x - x'\|^2}{2 \alpha \ell^2} + 1 \right) - \|x - x'\|^2}{2 \alpha \ell^2 + \|x - x'\|^2} \right)}

The derivative of \eqn{K_{\text{RQ}}(x, x')} with respect to \eqn{\ell} is given by:

\eqn{\frac{dK_{\text{RQ}}(x, x')}{d\ell} = \sigma^2 \cdot \left(1 + \frac{\|x - x'\|^2}{2 \alpha \ell^2}\right)^{-\alpha-1} \cdot \left( \frac{\|x - x'\|^2}{\ell^3} \right)}

\eqn{\frac{dK_{\text{RQ}}(x, x')}{d\ell} = K_{\text{RQ}}(x, x') \cdot \frac{\|x - x'\|^2}{\ell^3 \left(1 + \frac{\|x - x'\|^2}{2 \alpha \ell^2}\right)}}
}
