% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_kernel.R
\name{kernel_deriv,SEKernel-method}
\alias{kernel_deriv,SEKernel-method}
\title{Kernel Derivative Method for SEKernel}
\usage{
\S4method{kernel_deriv}{SEKernel}(obj, x, y, param)
}
\arguments{
\item{obj}{An object of class \code{SEKernel}.}

\item{x}{A matrix of input vectors.}

\item{y}{A matrix of input vectors.}

\item{param}{The parameter with respect to which the derivative is computed. Can be "variance_se" or "length_scale_se".}
}
\value{
A matrix representing the derivative of the kernel evaluations.
}
\description{
Computes the derivative of the kernel function with respect to a parameter for an \code{SEKernel} object.
Supports derivatives with respect to the variance and length scale parameters.

The derivative of \eqn{K_{\text{SE}}(x, x')} with respect to \eqn{\sigma^2} is given by:

\eqn{\frac{dK_{\text{SE}}(x, x')}{d\sigma^2} = \exp\left(-\frac{\|x - x'\|^2}{2\ell^2}\right)}

\eqn{\frac{dK_{\text{SE}}(x, x')}{d\sigma^2} = \frac{K_{\text{SE}}(x, x')}{\sigma^2}}

The derivative of \eqn{K_{\text{SE}}(x, x')} with respect to \eqn{\ell} is given by:

\eqn{\frac{dK_{\text{SE}}(x, x')}{d\ell} = \sigma^2 \exp\left(-\frac{\|x - x'\|^2}{2\ell^2}\right) \cdot \frac{\|x - x'\|^2}{\ell^3}}

\eqn{\frac{dK_{\text{SE}}(x, x')}{d\ell} = K_{\text{SE}}(x, x') \cdot \frac{\|x - x'\|^2}{\ell^3}}
}
