% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_kernel.R
\name{kernel_deriv,SumKernel-method}
\alias{kernel_deriv,SumKernel-method}
\title{Kernel Derivative Method for SumKernel}
\usage{
\S4method{kernel_deriv}{SumKernel}(obj, x, y, param)
}
\arguments{
\item{obj}{An object of class \code{SumKernel}.}

\item{x}{A matrix of input vectors.}

\item{y}{A matrix of input vectors.}

\item{param}{The parameter with respect to which the derivative is computed.}
}
\value{
A matrix representing the derivative of the kernel evaluations.
}
\description{
Computes the derivative of the kernel function with respect to a parameter for a \code{SumKernel} object.
The derivative is computed as the sum of the derivatives of the constituent kernels.
}
