% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_kernel.R
\name{kernel_deriv}
\alias{kernel_deriv}
\title{Generic Function for Kernel Derivative Computation}
\usage{
kernel_deriv(obj, x, y, param)
}
\arguments{
\item{obj}{An object inheriting from the \code{AbstractKernel} class.}

\item{x}{A matrix of input vectors.}

\item{y}{A matrix of input vectors.}

\item{param}{The parameter with respect to which the derivative is computed.}
}
\value{
A matrix representing the derivative of the kernel evaluations.
}
\description{
This generic function computes the derivative of the kernel function with respect to a parameter.
}
