% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dot-product-kernels.R
\name{linear_kernel}
\alias{linear_kernel}
\title{Linear kernel}
\usage{
linear_kernel(
  slope_var,
  slope_var_parametrisation = log_exp_parametrisation(),
  distance_function = dot_product,
  engine = dense_engine()
)
}
\arguments{
\item{slope_var}{A strictly positive numeric value. Controls the slope.}

\item{slope_var_parametrisation}{A \link{parametrisation} for \code{slope_var}.
Defaults to \code{\link[=log_exp_parametrisation]{log_exp_parametrisation()}}.}

\item{distance_function}{A distance function, or a shortcut name (see
\code{\link[=resolve_distance_function]{resolve_distance_function()}}). Defaults to \code{\link[=dot_product]{dot_product()}}.}

\item{engine}{A computation engine. Defaults to \code{\link[=dense_engine]{dense_engine()}}.}
}
\value{
An object of class \code{linear_kernel}/\code{dot_product_kernel}/\code{kernel}.
}
\description{
\deqn{k(x_1, x_2) = \mathrm{slope\_var} \cdot \langle x_1, x_2 \rangle}

Note: samples/posteriors from this kernel always cross the point
\verb{(0, 0)}. To allow other crossing points, either use \code{\link[=affine_kernel]{affine_kernel()}}
with a fixed \code{offset}, or add a \code{\link[=constant_kernel]{constant_kernel()}} to a
\code{linear_kernel()}, where the constant value represents the variance at
the crossing point.
}
\examples{
k <- linear_kernel(slope_var = 1)
evaluate(k, c(0, 1, 2))
}
