% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_kernel.R
\docType{class}
\name{LinearKernel-class}
\alias{LinearKernel-class}
\title{Linear Kernel Class}
\description{
A class representing a linear kernel, which inherits from the \code{AbstractKernel} class.
The linear kernel is defined by two variance parameters and a constant term.
}
\section{Slots}{

\describe{
\item{\code{sigma2_b}}{A numeric value representing the bias variance.}

\item{\code{sigma2_v}}{A numeric value representing the variance scaling factor.}

\item{\code{c}}{A numeric value representing the constant offset.}
}}

