% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_kernel.R
\docType{class}
\name{MaternKernel12-class}
\alias{MaternKernel12-class}
\title{Matern Kernel 1/2 Class}
\description{
A class representing a Matern kernel with smoothness parameter \eqn{\nu} = 1/2 , which inherits from the \code{AbstractKernel} class.
This kernel is a type of covariance function used in Gaussian process regression.
}
\section{Slots}{

\describe{
\item{\code{length_scale_mat}}{A numeric value representing the length scale parameter.}
}}

