% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_kernel.R
\docType{class}
\name{NoiseKernel-class}
\alias{NoiseKernel-class}
\title{Noise Kernel Class}
\description{
A class representing a noise kernel, which inherits from the \code{AbstractKernel} class.
This kernel models noise in the data by returning a diagonal matrix scaled by a constant value.
}
\section{Slots}{

\describe{
\item{\code{value_c}}{A numeric value representing the noise variance.}
}}

