% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_kernel.R
\name{pairwise_kernel,AbstractKernel-method}
\alias{pairwise_kernel,AbstractKernel-method}
\title{Pairwise Kernel Method for AbstractKernel}
\usage{
\S4method{pairwise_kernel}{AbstractKernel}(obj, x, y)
}
\arguments{
\item{obj}{An object of class \code{AbstractKernel}.}

\item{x}{A matrix of input vectors.}

\item{y}{A matrix of input vectors.}
}
\description{
This method is intended to compute the pairwise kernel matrix for an \code{AbstractKernel} object.
It must be implemented in subclasses.
}
