% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_kernel.R
\name{pairwise_kernel,ConstantKernel-method}
\alias{pairwise_kernel,ConstantKernel-method}
\title{Pairwise Kernel Method for ConstantKernel}
\usage{
\S4method{pairwise_kernel}{ConstantKernel}(obj, x, y)
}
\arguments{
\item{obj}{An object of class \code{ConstantKernel}.}

\item{x}{A matrix of input vectors.}

\item{y}{A matrix of input vectors.}
}
\value{
A matrix filled with the constant value of the kernel.
}
\description{
Computes the pairwise kernel matrix for a \code{ConstantKernel} object.

Let \eqn{c \in \mathbb{R}_{+}^{*}}. Consider the constant kernel defined by:

\eqn{K_{\text{cst}}(x, x') = c \cdot \begin{bmatrix}
1 & 1 & \cdots & 1 \\
1 & 1 & \cdots & 1 \\
\vdots & \vdots & \ddots & \vdots \\
1 & 1 & \cdots & 1
\end{bmatrix}}

The kernel matrix is filled with the constant value of the kernel.
}
