% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_kernel.R
\name{pairwise_kernel,LinearKernel-method}
\alias{pairwise_kernel,LinearKernel-method}
\title{Pairwise Kernel Method for LinearKernel}
\usage{
\S4method{pairwise_kernel}{LinearKernel}(obj, x, y)
}
\arguments{
\item{obj}{An object of class \code{LinearKernel}.}

\item{x}{A matrix of input vectors.}

\item{y}{A matrix of input vectors.}
}
\value{
A matrix representing the pairwise kernel evaluations.
}
\description{
Computes the pairwise kernel matrix for a \code{LinearKernel} object using the linear kernel formula.
Let \eqn{\sigma^2_b \in \mathbb{R}_{+}}, \eqn{\sigma^2_v \in \mathbb{R}_{+}}, and \eqn{c \in \mathbb{R}}.
Consider the linear kernel defined by:

\eqn{K_{\text{Lin}}(x, x') = \sigma^2_b + \sigma^2_v (x - c)(x' - c)}
}
