% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_kernel.R
\name{pairwise_kernel,MaternKernel12-method}
\alias{pairwise_kernel,MaternKernel12-method}
\title{Pairwise Kernel Method for MaternKernel12}
\usage{
\S4method{pairwise_kernel}{MaternKernel12}(obj, x, y)
}
\arguments{
\item{obj}{An object of class \code{MaternKernel12}.}

\item{x}{A matrix of input vectors.}

\item{y}{A matrix of input vectors.}
}
\value{
A matrix representing the pairwise kernel evaluations.
}
\description{
Computes the pairwise kernel matrix for a \code{MaternKernel12} object using the Matern kernel formula with \eqn{\nu = 1/2}.
Let \(\eqn{\ell \in \mathbb{R}_{+}^{*}}\) and \(\eqn{\sigma^2 \in \mathbb{R}_{+}}\). Consider the Matern Kernel with \eqn{\nu = 1/2} defined by:

\eqn{K_{\text{Mat}_{\nu = \frac{1}{2}}}(x, x') = \sigma^2 \cdot \exp\left(-\frac{\|x - x'\|^2}{\ell}\right)}
}
