% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_kernel.R
\name{pairwise_kernel,MaternKernel32-method}
\alias{pairwise_kernel,MaternKernel32-method}
\title{Pairwise Kernel Method for MaternKernel32}
\usage{
\S4method{pairwise_kernel}{MaternKernel32}(obj, x, y)
}
\arguments{
\item{obj}{An object of class \code{MaternKernel32}.}

\item{x}{A matrix of input vectors.}

\item{y}{A matrix of input vectors.}
}
\value{
A matrix representing the pairwise kernel evaluations.
}
\description{
Computes the pairwise kernel matrix for a \code{MaternKernel32} object using the Matern kernel formula with \eqn{\nu = 3/2}.
Let \eqn{\ell \in \mathbb{R}_{+}^{*}} and \eqn{\sigma^2 \in \mathbb{R}_{+}}. Consider the Matern Kernel with \eqn{\nu = 3/2} defined by:

\eqn{K_{\text{Mat}_{\nu = \frac{3}{2}}}(x, x') = \sigma^2 \cdot \left( \frac{\sqrt{3} \cdot \|x - x'\|^2}{\ell} + 1 \right) \exp\left(-\frac{\sqrt{3} \cdot \|x - x'\|^2}{\ell}\right)}
}
