% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_kernel.R
\name{pairwise_kernel,NoiseKernel-method}
\alias{pairwise_kernel,NoiseKernel-method}
\title{Pairwise Kernel Method for NoiseKernel}
\usage{
\S4method{pairwise_kernel}{NoiseKernel}(obj, x, y)
}
\arguments{
\item{obj}{An object of class \code{NoiseKernel}.}

\item{x}{A matrix of input vectors.}

\item{y}{A matrix of input vectors.}
}
\value{
A diagonal matrix scaled by the noise variance value.
}
\description{
Computes the pairwise kernel matrix for a \code{NoiseKernel} object.
Let \eqn{c \in \mathbb{R}_{+}^{*}}. Consider the noise kernel defined by:

\eqn{K_{\text{noise}}(x, x') = c \cdot I_n}

The kernel matrix is a diagonal matrix scaled by the noise variance value.
}
