% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_kernel.R
\name{pairwise_kernel,PeriodicKernel-method}
\alias{pairwise_kernel,PeriodicKernel-method}
\title{Pairwise Kernel Method for PeriodicKernel}
\usage{
\S4method{pairwise_kernel}{PeriodicKernel}(obj, x, y)
}
\arguments{
\item{obj}{An object of class \code{PeriodicKernel}.}

\item{x}{A matrix of input vectors.}

\item{y}{A matrix of input vectors.}
}
\value{
A matrix representing the pairwise kernel evaluations.
}
\description{
Computes the pairwise kernel matrix for a \code{PeriodicKernel} object using the periodic kernel formula.
Let \eqn{\sigma^2 \in \mathbb{R}_{+}}, \eqn{\ell \in \mathbb{R}}, and \eqn{p \in \mathbb{R}}.
Consider the periodic kernel defined by:

\eqn{K_{\text{Per}}(x, x') = \sigma^2 \exp \left( -\frac{2}{\ell^2} \sin^2 \left( \pi \frac{x - x'}{p} \right) \right)}
}
