% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_kernel.R
\name{pairwise_kernel,ProductKernel-method}
\alias{pairwise_kernel,ProductKernel-method}
\title{Pairwise Kernel Method for ProductKernel}
\usage{
\S4method{pairwise_kernel}{ProductKernel}(obj, x, y)
}
\arguments{
\item{obj}{An object of class \code{ProductKernel}.}

\item{x}{A matrix of input vectors.}

\item{y}{A matrix of input vectors.}
}
\value{
A matrix representing the product of the pairwise kernel evaluations of the constituent kernels.
}
\description{
Computes the pairwise kernel matrix for a \code{ProductKernel} object by multiplying the pairwise kernel matrices of its constituent kernels.
}
