% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_kernel.R
\name{pairwise_kernel,RationalQuadraticKernel-method}
\alias{pairwise_kernel,RationalQuadraticKernel-method}
\title{Pairwise Kernel Method for RationalQuadraticKernel}
\usage{
\S4method{pairwise_kernel}{RationalQuadraticKernel}(obj, x, y)
}
\arguments{
\item{obj}{An object of class \code{RationalQuadraticKernel}.}

\item{x}{A matrix of input vectors.}

\item{y}{A matrix of input vectors.}
}
\value{
A matrix representing the pairwise kernel evaluations.
}
\description{
Computes the pairwise kernel matrix for a \code{RationalQuadraticKernel} object using the Rational Quadratic kernel formula.
Let \eqn{\sigma^2 \in \mathbb{R}_{+}}, \eqn{\alpha \in \mathbb{R}_{+}^{*}}, and \eqn{\ell \in \mathbb{R}_{+}^{*}}.
Consider the RQ kernel defined by:

\eqn{K_{\text{RQ}}(x, x') = \sigma^2 \left(1 + \frac{\|x - x'\|^2}{2 \alpha \ell^2}\right)^{-\alpha}}
}
