% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_kernel.R
\name{pairwise_kernel,SEKernel-method}
\alias{pairwise_kernel,SEKernel-method}
\title{Pairwise Kernel Method for SEKernel}
\usage{
\S4method{pairwise_kernel}{SEKernel}(obj, x, y)
}
\arguments{
\item{obj}{An object of class \code{SEKernel}.}

\item{x}{A matrix of input vectors.}

\item{y}{A matrix of input vectors.}
}
\value{
A matrix representing the pairwise kernel evaluations.
}
\description{
Computes the pairwise kernel matrix for an \code{SEKernel} object using the squared exponential kernel formula.
Let \eqn{\sigma \in \mathbb{R}_{+}} and \eqn{\ell \in \mathbb{R}_{+}^{*}}.
The squared exponential kernel is defined by:

\eqn{K_{\text{SE}}(x, x') = \sigma^2 \exp\left(-\frac{\|x - x'\|^2}{2\ell^2}\right)},

where \eqn{\sigma} is the signal standard deviation and \eqn{\ell} is the length scale parameter.
}
