% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_kernel.R
\name{pairwise_kernel,SumKernel-method}
\alias{pairwise_kernel,SumKernel-method}
\title{Pairwise Kernel Method for SumKernel}
\usage{
\S4method{pairwise_kernel}{SumKernel}(obj, x, y)
}
\arguments{
\item{obj}{An object of class \code{SumKernel}.}

\item{x}{A matrix of input vectors.}

\item{y}{A matrix of input vectors.}
}
\value{
A matrix representing the sum of the pairwise kernel evaluations of the constituent kernels.
}
\description{
Computes the pairwise kernel matrix for a \code{SumKernel} object by summing the pairwise kernel matrices of its constituent kernels.
}
