% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_kernel.R
\name{pairwise_kernel}
\alias{pairwise_kernel}
\title{Generic Function for Pairwise Kernel Computation}
\usage{
pairwise_kernel(obj, x, y)
}
\arguments{
\item{obj}{An object inheriting from the \code{AbstractKernel} class.}

\item{x}{A matrix of input vectors.}

\item{y}{A matrix of input vectors.}
}
\value{
A matrix representing the pairwise kernel evaluations.
}
\description{
This generic function computes the pairwise kernel matrix between two sets of vectors.
}
