% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_kernel.R
\docType{class}
\name{PeriodicKernel-class}
\alias{PeriodicKernel-class}
\title{Periodic Kernel Class}
\description{
A class representing a Periodic kernel, which inherits from the \code{AbstractKernel} class.
The Periodic kernel is used to model periodic functions and is defined by a variance, length scale, and period.
}
\section{Slots}{

\describe{
\item{\code{variance_per}}{A numeric value representing the signal variance.}

\item{\code{length_scale_per}}{A numeric value representing the length scale parameter.}

\item{\code{period}}{A numeric value representing the period of the periodic function.}
}}

