% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_kernel_vs_distance}
\alias{plot_kernel_vs_distance}
\title{Plot Kernel Values vs Distance}
\usage{
plot_kernel_vs_distance(kernel, input_values)
}
\arguments{
\item{kernel}{A kernel object for which to compute the pairwise kernel values.}

\item{input_values}{A vector of input values for which to compute distances and kernel values.}
}
\value{
A ggplot object showing the kernel values versus distances.
}
\description{
Plots kernel values against the distance between input values.
}
