% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_kernel.R
\name{pretty_print,AbstractKernel-method}
\alias{pretty_print,AbstractKernel-method}
\title{Pretty Print Method for AbstractKernel}
\usage{
\S4method{pretty_print}{AbstractKernel}(obj)
}
\arguments{
\item{obj}{An object of class \code{AbstractKernel}.}
}
\value{
A string representation of the object.
}
\description{
This method provides a string representation for objects of class \code{AbstractKernel}.
}
