% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_kernel.R
\name{pretty_print,ConstantKernel-method}
\alias{pretty_print,ConstantKernel-method}
\title{Pretty Print Method for ConstantKernel}
\usage{
\S4method{pretty_print}{ConstantKernel}(obj)
}
\arguments{
\item{obj}{An object of class \code{ConstantKernel}.}
}
\value{
A string representation of the object.
}
\description{
Provides a string representation for objects of class \code{ConstantKernel}.
}
