% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_kernel.R
\name{pretty_print,LinearKernel-method}
\alias{pretty_print,LinearKernel-method}
\title{Pretty Print Method for LinearKernel}
\usage{
\S4method{pretty_print}{LinearKernel}(obj)
}
\arguments{
\item{obj}{An object of class \code{LinearKernel}.}
}
\value{
A string representation of the object.
}
\description{
Provides a string representation for objects of class \code{LinearKernel}.
}
