% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_kernel.R
\name{pretty_print,MaternKernel52-method}
\alias{pretty_print,MaternKernel52-method}
\title{Pretty Print Method for MaternKernel52}
\usage{
\S4method{pretty_print}{MaternKernel52}(obj)
}
\arguments{
\item{obj}{An object of class \code{MaternKernel52}.}
}
\value{
A string representation of the object.
}
\description{
Provides a string representation for objects of class \code{MaternKernel52}.
}
