% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_kernel.R
\name{pretty_print,NoiseKernel-method}
\alias{pretty_print,NoiseKernel-method}
\title{Pretty Print Method for NoiseKernel}
\usage{
\S4method{pretty_print}{NoiseKernel}(obj)
}
\arguments{
\item{obj}{An object of class \code{NoiseKernel}.}
}
\value{
A string representation of the object.
}
\description{
Provides a string representation for objects of class \code{NoiseKernel}.
}
