% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_kernel.R
\name{pretty_print,PeriodicKernel-method}
\alias{pretty_print,PeriodicKernel-method}
\title{Pretty Print Method for PeriodicKernel}
\usage{
\S4method{pretty_print}{PeriodicKernel}(obj)
}
\arguments{
\item{obj}{An object of class \code{PeriodicKernel}.}
}
\value{
A string representation of the object.
}
\description{
Provides a string representation for objects of class \code{PeriodicKernel}.
}
