% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_kernel.R
\name{pretty_print,ProductKernel-method}
\alias{pretty_print,ProductKernel-method}
\title{Pretty Print Method for ProductKernel}
\usage{
\S4method{pretty_print}{ProductKernel}(obj)
}
\arguments{
\item{obj}{An object of class \code{ProductKernel}.}
}
\value{
A string representation of the object.
}
\description{
Provides a string representation for objects of class \code{ProductKernel}.
}
