% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_kernel.R
\name{pretty_print,RationalQuadraticKernel-method}
\alias{pretty_print,RationalQuadraticKernel-method}
\title{Pretty Print Method for RationalQuadraticKernel}
\usage{
\S4method{pretty_print}{RationalQuadraticKernel}(obj)
}
\arguments{
\item{obj}{An object of class \code{RationalQuadraticKernel}.}
}
\value{
A string representation of the object.
}
\description{
Provides a string representation for objects of class \code{RationalQuadraticKernel}.
}
