% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_kernel.R
\name{pretty_print,SEKernel-method}
\alias{pretty_print,SEKernel-method}
\title{Pretty Print Method for SEKernel}
\usage{
\S4method{pretty_print}{SEKernel}(obj)
}
\arguments{
\item{obj}{An object of class \code{SEKernel}.}
}
\value{
A string representation of the object.
}
\description{
Provides a string representation for objects of class \code{SEKernel}.
}
