% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_kernel.R
\name{pretty_print,SumKernel-method}
\alias{pretty_print,SumKernel-method}
\title{Pretty Print Method for SumKernel}
\usage{
\S4method{pretty_print}{SumKernel}(obj)
}
\arguments{
\item{obj}{An object of class \code{SumKernel}.}
}
\value{
A string representation of the object.
}
\description{
Provides a string representation for objects of class \code{SumKernel}.
}
