% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_kernel.R
\name{pretty_print}
\alias{pretty_print}
\title{Generic Function for Pretty Printing}
\usage{
pretty_print(obj)
}
\arguments{
\item{obj}{An object inheriting from the \code{AbstractKernel} class.}
}
\value{
A string representation of the kernel.
}
\description{
This generic function provides a pretty-printed string representation of a kernel object.
}
