% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_kernel.R
\docType{class}
\name{ProductKernel-class}
\alias{ProductKernel-class}
\title{Product Kernel Class}
\description{
A class representing a product of kernels, which inherits from the \code{AbstractKernel} class.
This kernel allows combining multiple kernel objects by multiplying their outputs.
}
\section{Slots}{

\describe{
\item{\code{kernels}}{A list of kernel objects that are multiplied together.}
}}

