% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_kernel.R
\docType{class}
\name{RationalQuadraticKernel-class}
\alias{RationalQuadraticKernel-class}
\title{Rational Quadratic Kernel Class}
\description{
A class representing a Rational Quadratic kernel, which inherits from the \code{AbstractKernel} class.
The Rational Quadratic kernel is a scale mixture of Squared Exponential kernels with different characteristic length-scales.
}
\section{Slots}{

\describe{
\item{\code{variance_rq}}{A numeric value representing the signal variance.}

\item{\code{length_scale_rq}}{A numeric value representing the length scale parameter.}

\item{\code{alpha_rq}}{A numeric value representing the scale mixture parameter.}
}}

