% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_kernel.R
\docType{class}
\name{SEKernel-class}
\alias{SEKernel-class}
\title{Squared Exponential Kernel Class}
\description{
A class representing a Squared Exponential (SE) kernel, which inherits from the \code{AbstractKernel} class.
The SE kernel is commonly used in Gaussian processes and is defined by a variance and length scale parameter.
}
\section{Slots}{

\describe{
\item{\code{variance_se}}{A numeric value representing the signal variance.}

\item{\code{length_scale_se}}{A numeric value representing the length scale parameter.}
}}

