% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_kernel.R
\name{set_hyperparameters}
\alias{set_hyperparameters}
\title{Replace Hps Kernels}
\usage{
set_hyperparameters(kernel, values)
}
\arguments{
\item{kernel}{Previous kernel}

\item{values}{New HPs}
}
\value{
A kernel with new HPs
}
\description{
Update HPs of complex kernels
}
\examples{
# Assuming SEKernel is a subclass of AbstractKernel with hyperparameters
kernel <- new("SEKernel", variance_se = 1, length_scale_se = 1)
updated_kernel <- set_hyperparameters(kernel, c(variance_se = 2, length_scale_se = 0.5))

}
