% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_kernel.R
\name{show,AbstractKernel-method}
\alias{show,AbstractKernel-method}
\title{Show Method for AbstractKernel}
\usage{
\S4method{show}{AbstractKernel}(object)
}
\arguments{
\item{object}{An object of class \code{AbstractKernel}.}
}
\description{
This method provides a basic display for objects of class \code{AbstractKernel}.
}
