% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_kernel.R
\name{show,ConstantKernel-method}
\alias{show,ConstantKernel-method}
\title{Show Method for ConstantKernel}
\usage{
\S4method{show}{ConstantKernel}(object)
}
\arguments{
\item{object}{An object of class \code{ConstantKernel}.}
}
\description{
Provides a display for objects of class \code{ConstantKernel}.
}
