% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_kernel.R
\name{show,LinearKernel-method}
\alias{show,LinearKernel-method}
\title{Show Method for LinearKernel}
\usage{
\S4method{show}{LinearKernel}(object)
}
\arguments{
\item{object}{An object of class \code{LinearKernel}.}
}
\description{
Provides a display for objects of class \code{LinearKernel}.
}
