% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_kernel.R
\name{show,MaternKernel12-method}
\alias{show,MaternKernel12-method}
\title{Show Method for MaternKernel12}
\usage{
\S4method{show}{MaternKernel12}(object)
}
\arguments{
\item{object}{An object of class \code{MaternKernel12}.}
}
\description{
Provides a display for objects of class \code{MaternKernel12}.
}
