% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_kernel.R
\name{show,MaternKernel32-method}
\alias{show,MaternKernel32-method}
\title{Show Method for MaternKernel32}
\usage{
\S4method{show}{MaternKernel32}(object)
}
\arguments{
\item{object}{An object of class \code{MaternKernel32}.}
}
\description{
Provides a display for objects of class \code{MaternKernel32}.
}
