% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_kernel.R
\name{show,MaternKernel52-method}
\alias{show,MaternKernel52-method}
\title{Show Method for MaternKernel52}
\usage{
\S4method{show}{MaternKernel52}(object)
}
\arguments{
\item{object}{An object of class \code{MaternKernel52}.}
}
\description{
Provides a display for objects of class \code{MaternKernel52}.
}
