% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_kernel.R
\name{show,NoiseKernel-method}
\alias{show,NoiseKernel-method}
\title{Show Method for NoiseKernel}
\usage{
\S4method{show}{NoiseKernel}(object)
}
\arguments{
\item{object}{An object of class \code{NoiseKernel}.}
}
\description{
Provides a display for objects of class \code{NoiseKernel}.
}
