% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_kernel.R
\name{show,PeriodicKernel-method}
\alias{show,PeriodicKernel-method}
\title{Show Method for PeriodicKernel}
\usage{
\S4method{show}{PeriodicKernel}(object)
}
\arguments{
\item{object}{An object of class \code{PeriodicKernel}.}
}
\description{
Provides a display for objects of class \code{PeriodicKernel}.
}
