% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_kernel.R
\name{show,ProductKernel-method}
\alias{show,ProductKernel-method}
\title{Show Method for ProductKernel}
\usage{
\S4method{show}{ProductKernel}(object)
}
\arguments{
\item{object}{An object of class \code{ProductKernel}.}
}
\description{
Provides a display for objects of class \code{ProductKernel}.
}
