% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_kernel.R
\name{show,RationalQuadraticKernel-method}
\alias{show,RationalQuadraticKernel-method}
\title{Show Method for RationalQuadraticKernel}
\usage{
\S4method{show}{RationalQuadraticKernel}(object)
}
\arguments{
\item{object}{An object of class \code{RationalQuadraticKernel}.}
}
\description{
Provides a display for objects of class \code{RationalQuadraticKernel}.
}
