% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_kernel.R
\name{show,SEKernel-method}
\alias{show,SEKernel-method}
\title{Show Method for SEKernel}
\usage{
\S4method{show}{SEKernel}(object)
}
\arguments{
\item{object}{An object of class \code{SEKernel}.}
}
\description{
Provides a display for objects of class \code{SEKernel}.
}
