% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_kernel.R
\name{show,SumKernel-method}
\alias{show,SumKernel-method}
\title{Show Method for SumKernel}
\usage{
\S4method{show}{SumKernel}(object)
}
\arguments{
\item{object}{An object of class \code{SumKernel}.}
}
\description{
Provides a display for objects of class \code{SumKernel}.
}
