% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dot-product-kernels.R
\name{sigmoid_kernel}
\alias{sigmoid_kernel}
\title{Sigmoid kernel}
\usage{
sigmoid_kernel(
  alpha,
  constant,
  alpha_parametrisation = log_exp_parametrisation(),
  distance_function = dot_product,
  engine = dense_engine()
)
}
\arguments{
\item{alpha}{A strictly positive numeric value.}

\item{constant}{A numeric value (any real value).}

\item{alpha_parametrisation}{A \link{parametrisation} for \code{alpha}. Defaults to
\code{\link[=log_exp_parametrisation]{log_exp_parametrisation()}}.}

\item{distance_function}{A distance function, or a shortcut name (see
\code{\link[=resolve_distance_function]{resolve_distance_function()}}). Defaults to \code{\link[=dot_product]{dot_product()}}.}

\item{engine}{A computation engine. Defaults to \code{\link[=dense_engine]{dense_engine()}}.}
}
\value{
An object of class \code{sigmoid_kernel}/\code{dot_product_kernel}/\code{kernel}.
}
\description{
\deqn{k(x_1, x_2) = \tanh(\alpha \cdot \langle x_1, x_2 \rangle + c)}
}
\examples{
k <- sigmoid_kernel(alpha = 1, constant = 0)
evaluate(k, c(0, 1, 2))
}
