% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_kernel.R
\docType{class}
\name{SumKernel-class}
\alias{SumKernel-class}
\title{Sum Kernel Class}
\description{
A class representing a sum of kernels, which inherits from the \code{AbstractKernel} class.
This kernel allows combining multiple kernel objects by summing their outputs.
}
\section{Slots}{

\describe{
\item{\code{kernels}}{A list of kernel objects that are summed together.}
}}

