% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapped-param.R
\name{unwrap_param}
\alias{unwrap_param}
\title{Read the constrained value of a wrapped parameter}
\usage{
unwrap_param(x)
}
\arguments{
\item{x}{An object of class \code{wrapped_param}.}
}
\value{
The value in the natural (constrained) space.
}
\description{
Read the constrained value of a wrapped parameter
}
\examples{
unwrap_param(wrapped_param(2.5))
}
