% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other-kernels.R
\name{variance_kernel}
\alias{variance_kernel}
\title{Variance kernel}
\usage{
variance_kernel(
  variance = 1,
  variance_parametrisation = log_exp_parametrisation(),
  engine = dense_engine()
)
}
\arguments{
\item{variance}{A strictly positive numeric value. Defaults to \code{1}.}

\item{variance_parametrisation}{A \link{parametrisation} for \code{variance}.
Defaults to \code{\link[=log_exp_parametrisation]{log_exp_parametrisation()}}.}

\item{engine}{A computation engine. Defaults to \code{\link[=dense_engine]{dense_engine()}}.}
}
\value{
An object of class \code{variance_kernel}/\code{kernel}.
}
\description{
Returns the same strictly positive constant value everywhere. Typically
multiplied with another kernel (via \code{*}, see \link{kernel_arithmetic}) to
scale its output.

\deqn{k(x_1, x_2) = \mathrm{variance}}
}
\examples{
k <- variance_kernel(variance = 2)
evaluate(k, c(0, 1, 2))
}
