% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{visualize_kernel}
\alias{visualize_kernel}
\title{Visualize a Kernel Matrix}
\usage{
visualize_kernel(kernel, x)
}
\arguments{
\item{kernel}{A kernel object for which to compute the pairwise kernel matrix.}

\item{x}{A matrix or data frame where each row represents an input data point.}
}
\value{
A ggplot object displaying the kernel matrix.
}
\description{
Plots a heatmap of the kernel matrix computed from input data.
}
