% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optim_HPs.R
\name{optim_hp_L_BFGS_B}
\alias{optim_hp_L_BFGS_B}
\title{Optimize Hyperparameters for a kernel}
\usage{
optim_hp_L_BFGS_B(
  hp,
  db,
  mean,
  kern,
  post_cov,
  pen_diag = 1e-06,
  verbose = FALSE
)
}
\arguments{
\item{hp}{A vector of initial hyperparameters to be optimized.}

\item{db}{The dataset used for optimization.}

\item{mean}{The mean function}

\item{kern}{The kernel function}

\item{post_cov}{The posterior covariance function.}

\item{pen_diag}{A penalty term added to the diagonal of the covariance matrix for numerical stability.}

\item{verbose}{A logical value indicating whether to return the full optimization result or just the optimized parameters. Defaults to FALSE.}
}
\value{
If \code{verbose} is FALSE, a vector of optimized hyperparameters; otherwise, the full result from the \code{optim} function.
}
\description{
This function optimizes hyperparameters using the L-BFGS-B optimization method.
}
